lab-mqtt/requirements.txt
paho-mqtt==1.6.1
PyYAML==6.0.2
apscheduler==3.10.4
fastapi==0.115.0
uvicorn[standard]==0.30.6
Jinja2==3.1.4
